# Fix #3: Optimize memory access patterns for HBM2 on GFX906