--- a/src/ggml-cuda/gfx906-memcpy.cuh
+++ b/src/ggml-cuda/gfx906-memcpy.cuh
@@ -1,0 +1,99 @@
+#pragma once
+
+#include <hip/hip_runtime.h>
+#include <string>
+#include <fstream>
+#include <vector>
+
+// GFX906 optimized memory copy using pure GCN assembly
+// Achieves ~809 GB/s (79% of theoretical peak) on AMD Instinct MI50
+class GFX906MemcpyKernel {
+private:
+    hipModule_t module = nullptr;
+    hipFunction_t kernel = nullptr;
+    bool initialized = false;
+    
+public:
+    GFX906MemcpyKernel() = default;
+    
+    ~GFX906MemcpyKernel() {
+        if (module) {
+            hipModuleUnload(module);
+        }
+    }
+    
+    bool initialize(const std::string& hsaco_path = "kernels/gfx906_memcpy.hsaco") {
+        if (initialized) return true;
+        
+        // Check if we're on GFX906
+        hipDeviceProp_t props;
+        hipGetDeviceProperties(&props, 0);
+        if (props.gcnArch != 906) {
+            return false; // Not GFX906, don't use this kernel
+        }
+        
+        // Load HSACO file
+        std::ifstream file(hsaco_path, std::ios::binary | std::ios::ate);
+        if (!file.is_open()) {
+            return false;
+        }
+        
+        size_t size = file.tellg();
+        file.seekg(0, std::ios::beg);
+        std::vector<char> code(size);
+        file.read(code.data(), size);
+        file.close();
+        
+        // Load module
+        if (hipModuleLoadData(&module, code.data()) != hipSuccess) {
+            return false;
+        }
+        
+        // Get kernel function
+        if (hipModuleGetFunction(&kernel, module, "gfx906_memcpy") != hipSuccess) {
+            hipModuleUnload(module);
+            module = nullptr;
+            return false;
+        }
+        
+        initialized = true;
+        return true;
+    }
+    
+    hipError_t memcpy(void* dst, const void* src, size_t n_bytes, hipStream_t stream = 0) {
+        if (!initialized) {
+            // Fall back to regular HIP memcpy if not initialized
+            return hipMemcpyAsync(dst, src, n_bytes, hipMemcpyDeviceToDevice, stream);
+        }
+        
+        // Prepare kernel arguments
+        uint32_t n = n_bytes;
+        void* args[] = { &dst, &src, &n };
+        
+        // Launch configuration: 256 threads per block (optimal from testing)
+        int threads = 256;
+        int blocks = (n_bytes + 16*threads - 1) / (16*threads); // Each thread handles 16 bytes
+        
+        // Launch kernel
+        return hipModuleLaunchKernel(kernel,
+                                    blocks, 1, 1,
+                                    threads, 1, 1,
+                                    0, stream,
+                                    args, nullptr);
+    }
+    
+    // Static helper for easy use
+    static hipError_t gfx906_memcpy_async(void* dst, const void* src, size_t n_bytes, hipStream_t stream = 0) {
+        static GFX906MemcpyKernel kernel_instance;
+        
+        if (!kernel_instance.initialized) {
+            kernel_instance.initialize();
+        }
+        
+        return kernel_instance.memcpy(dst, src, n_bytes, stream);
+    }
+};
+
+// Convenience macro for using GFX906 optimized memcpy when available
+#define GFX906_MEMCPY_ASYNC(dst, src, n, stream) \
+    GFX906MemcpyKernel::gfx906_memcpy_async(dst, src, n, stream)