… docs/gfx906/memory-optimizations.md
… src/ggml-cuda/CMakeLists_gfx906.txt
… src/ggml-cuda/kernels/gfx906_memcpy.hsaco
… src/ggml-cuda/kernels/gfx906_memcpy.s
… src/ggml-cuda/test-gfx906-memory.cu