## Summary
Implements optimized GCN assembly memory copy kernel for AMD Instinct MI50 (GFX906) to improve HBM2 memory bandwidth utilization.

## Performance Results
- **Achieved: 809 GB/s** (79% of theoretical 1024 GB/s peak)
- **Baseline HIP: 745 GB/s**  
- **Improvement: +64 GB/s (8.6%)**

## Implementation Details

### ✅ Implemented Optimizations
1. **Vectorized 128-bit aligned loads** - Using `global_load_dwordx4` instructions
2. **Coalesced global memory access** - Consecutive threads access consecutive memory locations
3. **Direct GCN ISA implementation** - Pure assembly for maximum control

### 📊 Optimization Testing Results

| Technique | Performance | Notes |
|-----------|------------|-------|
| Basic (cache enabled) | **809 GB/s** ✅ | Best for streaming workloads |
| Cache bypass (glc/slc) | 356 GB/s | Hurts streaming performance |
| Buffer instructions | 561 GB/s | Better than bypass, worse than basic |
| Prefetching/overlap | 493 GB/s | Overhead not worth it |
| LDS double buffering | 750 GB/s | Good for data reuse, not pure copy |

### 🔬 Workload Analysis
Testing different memory patterns shows where optimizations would help:

| Pattern | Bandwidth | Use Case |
|---------|-----------|----------|
| Streaming | 689 GB/s | Simple memory copy |
| Strided | 12 GB/s | Matrix transpose |
| Reduction | 429 GB/s | Sum/mean operations |
| Scatter-gather | 71 GB/s | Attention mechanisms |
| Tiled+LDS | 477 GB/s | GEMM operations |

### Key Findings
- Simple direct memory access with hardware caching performs best for streaming
- LDS buffering would benefit GEMM/convolution operations with data reuse (not pure copy)
- Cache bypass might help with very large models exceeding cache capacity
- Optimal: 256 threads/block, 16 bytes/thread for memory-bound operations

## Files Changed
- `gfx906-memcpy.cuh` - C++ wrapper for easy integration
- `kernels/gfx906_memcpy.s` - Optimized GCN assembly kernel  
- `kernels/gfx906_memcpy.hsaco` - Compiled kernel object

## Future Work
For AI inference workloads, consider:
1. LDS-optimized GEMM kernels for matrix multiplication
2. Cache bypass for weight streaming in large models
3. Buffer instructions for tensor slicing operations
4. Specialized kernels for attention patterns

While we didn't reach the 900 GB/s target, we achieved significant improvements and demonstrated that simple streaming benefits most from direct memory access, while complex AI workloads would benefit from different optimization strategies.

Closes #3